[FIX]: Profile page only appears when logged in [REFACTOR]: Added the .sln file to the gitignore [REFACTOR]: Created useSupabase component to allow easier access to user login status. Refactored header and sidebar components to use it
