+++ .gitignore
@@ -5,6 +5,7 @@
 
 node_modules/**/node_modules
 node_modules/
+IoTM-Dashboard.sln
 
 # Logs
 logs
