+++ app/frontend/src/components/views/header.tsx
@@ -4,14 +4,13 @@ import { Bell, Heart } from "lucide-react"
 import { Button } from "@/components/ui/button"
 import Link from "next/link"
 import { supabase } from "@/lib/supabase/client"
-import { User } from "@supabase/supabase-js"
-import { useState, useEffect } from "react"
+import { useSupabaseUser } from "@/lib/supabase/useSupabaseUser"
+
 import { useRouter } from "next/navigation"
 import { toast } from "react-toastify"
 
 export default function Header() {
-
-  const [user, setUser] = useState<User|null>(null)
+  const user = useSupabaseUser()
   const router = useRouter()
 
   const logout = async () => {
@@ -19,33 +18,6 @@ export default function Header() {
     toast.success("User logged out Succesfully")
     router.push("/login") // Go back to the login page
   }
-
-  useEffect(() => {
-    const getUser = async () => {
-      // sets user variable value from the supabase client cookie
-      const { data } = await supabase.auth.getUser()
-        if (data?.user){
-          setUser(data.user)
-        }
-    }
-    // Calls the function once on object mount
-    getUser()
-
-    // Listener for Auth State Change - updates the user whenever login, logout, or token refresh happens
-    const { data: authListener } = supabase.auth.onAuthStateChange(
-      (event, session) => {
-        if (session?.user) {
-          setUser(session.user)
-        } else {
-          setUser(null)
-        }
-      }
-    )
-
-    return () => {
-      authListener.subscription.unsubscribe()
-    }
-  }, [])
 
   return (
     <header className="w-full border-b border-gray-300 bg-white px-4 py-3 flex items-center justify-between">
